requirements.txt
# BTC Elite Trader Dependencies

# Core dependencies (existing)
numpy>=1.26.0
pandas>=2.0.0
requests>=2.31.0
tenacity>=8.2.0
PyYAML>=6.0

# Exchange integration
ccxt>=4.0.0

# Async support
aiohttp>=3.9.0

# Telegram bot
python-telegram-bot>=20.0

# Database
asyncpg>=0.29.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# ML Regime Detection (optional - install for ML features)
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Type checking (optional)
# mypy>=1.0.0
# types-PyYAML
# types-requests
